requirements.txt
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.4
argparse==1.2.1
distribute==0.6.27
itsdangerous==0.23
plivo==0.9.1
python-dateutil==2.1
python-simple-hipchat==0.1
pytz==2013d
redis==2.8.0
requests==1.2.3
rq==0.3.11
simplejson==3.3.0
six==1.4.1
times==0.6.2
wsgiref==0.1.2
